further refactoring & updates
