+++ package.json
@@ -78,11 +78,11 @@
     "@typescript-eslint/eslint-plugin": "5.38.1",
     "@typescript-eslint/parser": "5.38.1",
     "@vscode/test-electron": "^2.1.5",
-    "@wdio/cli": "^7.24.0",
-    "@wdio/devtools-service": "^7.25.0",
-    "@wdio/local-runner": "^7.24.0",
-    "@wdio/mocha-framework": "^7.24.0",
-    "@wdio/spec-reporter": "^7.24.0",
+    "@wdio/cli": "^7.27.0",
+    "@wdio/devtools-service": "^7.27.0",
+    "@wdio/local-runner": "^7.27.0",
+    "@wdio/mocha-framework": "^7.26.0",
+    "@wdio/spec-reporter": "^7.26.0",
     "@yarnpkg/pnp": "^3.1.1-rc.12",
     "cypress": "^10.2.0",
     "esbuild": "^0.14.54",
@@ -107,7 +107,7 @@
     "universal-analytics": "0.5.3",
     "uuid": "^8.3.2",
     "vsce": "^2.6.3",
-    "wdio-vscode-service": "^4.2.0",
+    "wdio-vscode-service": "^4.2.1",
     "webdriverio": "^7.26.0",
     "webpack": "^5.47.0",
     "zone.js": "0.11.4"
